server: Renaming columns. Testing needed
